# doc-files/kypo-adaptive-training-swagger-open-api.yaml
...
swagger: "2.0"
info:
  version: "Version: 1.0.45"
  title: "KYPO Adaptive Training Service - API Reference"
host: "localhost:8082"
basePath: "/kypo-adaptive-training/api/v1"
...
schemes:
- "http"
- "https"
paths:
...
  /training-definitions/{definitionId}/phases/{phaseId}/tasks:
    post:
      tags:
      - "Tasks"
      summary: "Create a new task in a phase"
      description: "Creates a new default task in a specified training phase"
      operationId: "createTask"
      produces:
      - "application/json"
      parameters:
      - name: "phaseId"
        in: "path"
        description: "Training phase ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/TaskDTO"
        201:
          description: "Task created"
        500:
          description: "Unexpected application error"
      security:
      - bearerAuth: []
  /training-definitions/{definitionId}/phases/{phaseId}/tasks/{taskIdFrom}/move-to/{newPosition}:
    put:
      tags:
      - "Tasks"
      summary: "Move task to specified order"
      description: ""
      operationId: "moveTaskToSpecifiedOrder"
      produces:
      - "application/json"
      parameters:
      - name: "taskIdFrom"
        in: "path"
        description: "Task ID - from"
        required: true
        type: "integer"
        format: "int64"
      - name: "newPosition"
        in: "path"
        description: "Position (order) to which the task should be moved"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Task moved to specified order"
        500:
          description: "Unexpected application error"
      security:
      - bearerAuth: []
  /training-definitions/{definitionId}/phases/{phaseId}/tasks/{taskId}:
    get:
      tags:
      - "Tasks"
      summary: "Get task"
      description: "Get task detail associated with the specified training phase"
      operationId: "getTask"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "Task ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Task returned"
          schema:
            $ref: "#/definitions/TaskDTO"
        500:
          description: "Unexpected application error"
      security:
      - bearerAuth: []
    post:
      tags:
      - "Tasks"
      summary: "Clone task inside of the training phase"
      description: "Creates a new task with the same properties as the specified task\
        \ (pattern)"
      operationId: "cloneTask"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "Task ID"
        required: true
        type: "integer"
        format: "int64"
      - name: "phaseId"
        in: "path"
        description: "Training phase ID"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/TaskCopyDTO"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/TaskDTO"
        201:
          description: "Task cloned"
        500:
          description: "Unexpected application error"
      security:
      - bearerAuth: []
    put:
      tags:
      - "Tasks"
      summary: "Update task"
      description: "Update the specified task"
      operationId: "updateTask"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "Task ID"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Task to be updated"
        required: false
        schema:
          $ref: "#/definitions/TaskUpdateDTO"
      responses:
        200:
          description: "Task updated"
          schema:
            $ref: "#/definitions/TaskDTO"
        500:
          description: "Unexpected application error"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "Tasks"
      summary: "Remove a task"
      description: "Remove the specified task"
      operationId: "removeTask"
      produces:
      - "application/json"
      parameters:
      - name: "taskId"
        in: "path"
        description: "Task ID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Task removed"
          schema:
            $ref: "#/definitions/TaskDTO"
        500:
          description: "Unexpected application error"
      security:
      - bearerAuth: []
...
securityDefinitions:
  bearerAuth:
    description: "Obtain bearer token from OIDC provider and enter the token in the\
      \ following field. Use format 'Bearer {token}'"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  TaskCopyDTO:
    type: "object"
    required:
    - "answer"
    - "content"
    - "incorrectAnswerLimit"
    - "solution"
    - "title"
    properties:
      title:
        type: "string"
        example: "Task title"
        description: "Short description of task"
      content:
        type: "string"
        example: "Capture the flag"
        position: 1
        description: "The information that are displayed to a player"
      answer:
        type: "string"
        example: "secretFlag"
        position: 2
        description: "Keyword that must be found in the task. Necessary in order to\
          \ get to the next phase"
      solution:
        type: "string"
        example: "Open secret.txt"
        position: 3
        description: "Description how to get the answer"
      incorrectAnswerLimit:
        type: "integer"
        format: "int32"
        example: 5
        position: 4
        description: "It defines the allowed number of incorrect answers submitted\
          \ by the player"
        minimum: 0
        maximum: 100
      modifySandbox:
        type: "boolean"
        example: true
        position: 5
        description: "It defines whether the sandbox can be modified"
      sandboxChangeExpectedDuration:
        type: "integer"
        format: "int32"
        example: 15
        position: 6
        description: "It defines the expected duration of sandbox change defined in\
          \ seconds"
        minimum: 0
  TaskDTO:
    type: "object"
    required:
    - "answer"
    - "content"
    - "id"
    - "incorrectAnswerLimit"
    - "order"
    - "solution"
    - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
        description: "ID of task"
      title:
        type: "string"
        example: "Task title"
        description: "Short description of task"
      order:
        type: "integer"
        format: "int32"
        example: 1
        description: "Order of task in the training phase"
      content:
        type: "string"
        example: "Capture the flag"
        description: "The information that are displayed to a player"
      answer:
        type: "string"
        example: "secretFlag"
        description: "Keyword that must be found in the task. Necessary in order to\
          \ get to the next phase"
      solution:
        type: "string"
        example: "Open secret.txt"
        description: "Description how to get the answer"
      incorrectAnswerLimit:
        type: "integer"
        format: "int32"
        example: 5
        description: "It defines the allowed number of incorrect answers submitted\
          \ by the player"
      modifySandbox:
        type: "boolean"
        example: true
        description: "It defines whether the sandbox can be modified"
      sandboxChangeExpectedDuration:
        type: "integer"
        format: "int32"
        example: 15
        description: "It defines the expected duration of sandbox change defined in\
          \ seconds"
...
  TaskUpdateDTO:
    type: "object"
    required:
    - "answer"
    - "content"
    - "id"
    - "incorrectAnswerLimit"
    - "solution"
    - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
        example: 1
        description: "Main identifier of the task"
      title:
        type: "string"
        example: "Task title"
        description: "Short description of task"
      content:
        type: "string"
        example: "Capture the flag"
        position: 1
        description: "The information that are displayed to a player"
      sandboxChangeExpectedDuration:
        type: "integer"
        format: "int32"
        example: 15
        position: 1
        description: "It defines the expected duration of sandbox change defined in\
          \ seconds"
        minimum: 0
      answer:
        type: "string"
        example: "secretFlag"
        position: 2
        description: "Keyword that must be found in the task. Necessary in order to\
          \ get to the next phase"
        minLength: 0
        maxLength: 50
      solution:
        type: "string"
        example: "Open secret.txt"
        position: 3
        description: "Description how to get the answer"
      incorrectAnswerLimit:
        type: "integer"
        format: "int32"
        example: 5
        position: 4
        description: "It defines the allowed number of incorrect answers submitted\
          \ by the player"
        minimum: 0
        maximum: 100
      modifySandbox:
        type: "boolean"
        example: true
        position: 5
        description: "It defines whether the sandbox can be modified"
...